+++ c/src/minr.c
@@ -103,9 +103,9 @@ int main(int argc, char *argv[])
  * Read in 3D lattice descriptor:
  */
 
-  num_read = fread(&lat->xvoxels, sizeof(long), 1, infile);
-  num_read = fread(&lat->yvoxels, sizeof(long), 1, infile);
-  num_read = fread(&lat->zvoxels, sizeof(long), 1, infile);
+  num_read = fread(&lat->xvoxels, sizeof(uint32_t), 1, infile);
+  num_read = fread(&lat->yvoxels, sizeof(uint32_t), 1, infile);
+  num_read = fread(&lat->zvoxels, sizeof(uint32_t), 1, infile);
   num_read = fread(&lat->xbound, sizeof(struct bounds), 1, infile);
   num_read = fread(&lat->ybound, sizeof(struct bounds), 1, infile);
   num_read = fread(&lat->zbound, sizeof(struct bounds), 1, infile);
